third_party/WebKit/LayoutTests/http/tests/webfont/fontfaceset-status-attribute-expected.txt

Test for FontFaceSet.status attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.fonts.check('10px TestFont') is false
PASS document.fonts.status is "loaded"
PASS document.fonts.status is "loading"
PASS document.fonts.check('10px TestFont') is true
PASS document.fonts.status is "loaded"
PASS document.fonts.status is "loaded"
PASS successfullyParsed is true

TEST COMPLETE
